+++ crates/merman-render/src/lib.rs
@@ -176,13 +176,13 @@ pub fn layout_parsed_render_layout_only(
                 options.text_measurer.as_ref(),
             )?),
         ),
-        (RenderSemanticModel::Sequence(model), "sequence") => Ok(LayoutDiagram::SequenceDiagram(
-            sequence::layout_sequence_diagram_typed(
+        (RenderSemanticModel::Sequence(model), "sequence" | "zenuml") => Ok(
+            LayoutDiagram::SequenceDiagram(sequence::layout_sequence_diagram_typed(
                 model,
                 effective_config,
                 options.text_measurer.as_ref(),
-            )?,
-        )),
+            )?),
+        ),
         (RenderSemanticModel::Class(model), "classDiagram" | "class") => Ok(
             LayoutDiagram::ClassDiagramV2(class::layout_class_diagram_v2_typed_with_config(
                 model,